pyproject.toml
[project]
name = "discord-voice-calc"
version = "0.1.0"
description = ""
authors = [
    {name = "drago-suzuki58",email = "dorasuzu58@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)",
    "discord-py (>=2.4.0,<3.0.0)",
    "sqlalchemy (>=2.0.37,<3.0.0)",
    "aiosqlite (>=0.20.0,<0.21.0)",
    "loguru (>=0.7.3,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
